requirements.txt
markdown==3.5.2
PyYAML==6.0.1
Jinja2==3.1.3
Pygments==2.17.2
pymdown-extensions==10.7
beautifulsoup4==4.12.3
csscompressor==0.9.5